Reorg the fixtures
